Use a dedicated class for the "features" parameter
We cannot use Map, as URIs are not valid XML element names (which a Map would require
for maven to inject it).
Using Properties would in theory work, but looks ugly: "features" with child nodes called "property".

The 'features' integration test verifies that the generator is correctly initialized by using a custom
generator (compiled as part of the IT) that fails if the features aren't set with the proper values.
The verify.groovy script checks that the generator was used and left the marker file, otherwise it is assumed
that something silently broke the build.