Add support for specifying types to be generated explicitly
The types can be given in the <types> parameter (one <type> per type). By default
these get added to the list of types to generate. If <noImplicitTypes> is additionally
set to true then only the requested types will be generated.

While there: warn when there are no types to be generated: it's not a problem per se (we
simply don't have to do anything), but likely it is a misconfiguration of the plugin.